requirements.txt
speedtest-cli=2.1.2
python-dotenv==0.17.0
PyMySQL==1.0.2
SQLAlchemy==1.4.8